% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootree.R
\name{bootree}
\alias{bootree}
\title{Bootstrap Confidence Intervals for tree biomass samples}
\usage{
bootree(plotdata, samplename = "fragpar", area = 100, varnames = c(dbh =
  "diameter.breast.height", H = "plant.height", wd = "wood.density"),
  alom = chave, quantil = c(0.025, 0.975), nsim = 1000,
  convtoMgha = TRUE)
}
\arguments{
\item{plotdata}{a data frame with a forest sample plots data.}

\item{samplename}{character vector with the name of the column in plotdata that indicate the plots codes.}

\item{area}{numeric, plot size in square meters.}

\item{varnames}{vector of names variables variable diameter at breast hight, plant height and wood density, note that they should be  named as 'dbh', 'H' and 'wd' respectivelly. Put only those varibles that are in plotdata and needed for alometric calculation.}

\item{alom}{alometric function to be used, possibilities are \strong{chave} \strong{burgdeli}, \strong{scatena} and \strong{tiopolo}. Those are functions so should be write without quotes.}

\item{quantil}{Bootstrap quantiles to return, default is 95\% limite confidence interval (0.25 and 0.975 quatiles);}

\item{nsim}{numeric number of resamples from data to calculate bootstrap.}

\item{convtoMgha}{logical, if true convert AGB to Mg (tons) per hectar, otherwise the variable is in Kg per sample size.}
}
\value{
biomass calculation and bootstrap confidence interval, as well as standard deviation for the simulated bootstrap distribution
}
\description{
\code{bootree} return confidence interval for biomass (AGB) from samples of trees in a plot
}
\details{
Bootstrap resamples above ground biomass of each sample, reordering the sample tag.
}
\examples{

\dontrun{
bio = data.frame(dbh = rnorm(10, 10, 2), H= runif(10, 2, 8),
                 wd = rep(0.52,10), plots = rep(c(1,2),5))
bootree(bio, samplename = "plot", varnames =c(dbh="dbh", H = "H", wd = "wd"), nsim = 10)
metaEr(cl=10, rw=10, f0=0.2, pi=0.2, ce=0.15, tmax=100)
metaCiEr(cl=10, rw=10, f0=0.2, ci=0.2, ce=0.15, tmax=100)
}
}
\references{
SCATENA, F.N., SILVER, W., SICCAMA, T., JOHNSON, A. & SÁNCHEZ, M.J.  1993.
Biomass  and  nutrient  content  of  the  Bisley  Experimental Watersheds,
Luquillo Experimental Forest, Puerto Rico, before and after Hurricane Hugo, 1989.
Biotropica 25(1):15-27.

TIEPOLO,  G.,  CALMON,  M.  &  FERETTI,  A.R.  2002.  Measuring  and Monitoring
Carbon  Stocks  at  the  Guaraqueçaba  Climate  Action Project, Paraná, Brazil.
In: International Symposium on Forest Carbon Sequestration and Monitoring.
Extension Serie Taiwan Forestry Research Institute 153:98-115.

BURGER   D.M.;   DELITTI,   W.B.C.   allometric   models   for   estimating   the
phytomass of a secondary  Atlantic  Forest  area  of  southeastern  Brazil. 
Biota  Neotrop.8 (4)
}
\seealso{
\code{\link{chave}} \code{\link{scatena}} \code{\link{tiepolo}} \code{\link{burgdeli}}
}
\author{
Alexandre Adalardo de Oliveira and Paulo Inacio Prado
\email{aleadalardo@gmail.com}
}
