% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckPlotsOverlap.R
\name{CheckPlotsOverlap}
\alias{CheckPlotsOverlap}
\title{Checks if plots in a cartesian arena overlap}
\usage{
CheckPlotsOverlap(x_coords, y_coords, angles = NULL, shape = c("circle",
  "rectangle"), size = c(10, 10), distance = 0)
}
\arguments{
\item{x_coords}{vector of x coordinates of the plots center.}

\item{y_coords}{vector of x coordinates of the plots center.}

\item{angles}{vector of the orientation angles of the plots center.}

\item{shape}{a character string. Two shapes are possible: "rectangle" and "circle"}

\item{size}{numeric vector. For ''rectangle'' the dimension of x and y  sidesand for ''circle'' the radius size.}

\item{distance}{numeric value. The minimum distance among plots}
}
\value{
A list of the plots with their overlapping plots
}
\details{

}
\examples{
\dontrun{

x <- sample(seq(0, 1000), 50)
y <- sample(seq(0, 1000) , 50)
radius <- 10
overlap <- CheckPlotsOverlap(x, y, shape = "circle", size = radius, distance = 10)
}

}
