% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{dbh2barea}
\alias{dbh2barea}
\alias{barea2dbh}
\title{Internal function for samplePlotR package}
\usage{
dbh2barea(x, mscale = c("mm", "cm"))
}
\arguments{
\item{x}{a numeric vector.}

\item{mscale}{a character indicating the mesured scale (mm or cm).}

\item{statistics}{a function that calculates the statistics of interest from the dataframe.
The first argument should be the dataframe with the data and preferably should}
}
\value{
a vector of the \code{statistics}
}
\description{
Functions to calculate the parameters from the
sample.
}
\section{Details}{


Basic function to calculate descriptive statistics from a sample of a forest inventory. Basal area is calculate in 'cm' and the parameter \code{mscale} is the scale of the imput data.
}

\section{References}{

}

\examples{

\dontrun{
dbh2barea(x=seq(10:100, by=10), mscale= "mm")
}
}
