% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractSampleIndices.R
\name{ExtractSampleIndices}
\alias{ExtractSampleIndices}
\title{Extract biodiversity indices from sampled plots}
\usage{
ExtractSampleIndices(data, sampled_area = NA, species = "species",
  family = "family", biomass = "biomass", sample = "sample")
}
\arguments{
\item{data}{dataframe with a forest plot census data.}

\item{sampled_area}{numeric value. Total sampled area.}

\item{species}{character string with the name of the species variable.}

\item{family}{character string with the name of the family variable.}

\item{biomass}{character string with the name of the biomass variable.}

\item{sample}{character string with the name of the sample variable got from the rsample function.}
}
\value{
A data frame with the calculated biodiversity indices
}
\details{

}
\examples{
\dontrun{





}

}
