% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsample.R
\name{rsample}
\alias{rsample}
\title{Random sample elements in a cartesian arena}
\usage{
rsample(plotdata, dimx = c(0, 1000), dimy = c(0, 500), xname = "gx",
  yname = "gy", nsample = 100, shape = c("rectangle", "circle"),
  size = c(10, 10), maxiter = 10^5)
}
\arguments{
\item{plotdata}{a data frame with a forest plot census data.}

\item{dimx}{numeric vector with two positions minimum and  maximum value of the x coordinate.}

\item{dimy}{numeric vector with two positions minimum maximum value of the y coordinate.}

\item{xname}{character string with the name of the x coordinate variable.}

\item{yname}{character string with the name of the y coordinate variable.}

\item{nsample}{numeric value. The number of sample units.}

\item{shape}{a character string. Two shapes are possible: "rectangle" and "circle"}

\item{size}{numeric vector. For ''rectangle'' the dimension of x and y  sidesand for ''circle'' the radius size.}

\item{maxiter}{numeric value indicating the maximum number of trails to sample units without overlaping.}
}
\value{
a data frame with the same columns as plotdata plus a 'sample' representing the sample unit number that the element are in.
}
\description{
Draw a sample from data with elements in a cartesian coordinate system
}
\section{Details}{


Draw sample from elements in a cartesian coordinate system and return the subset of data in the sample.
}

\section{References}{

}

\examples{

\dontrun{
data(bci7)
plotData <- singleStem(bci7, nsample = 100, shape = "circle", size = 10)
}
}
