% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleStem.R
\name{singleStem}
\alias{singleStem}
\title{Calculate dbh for a tree with multiple stems}
\usage{
singleStem(plotdata, tag = "tag", dbh = "dbh", mscale = "mm")
}
\arguments{
\item{plotdata}{a data frame with a forest plot census data.}

\item{tag}{a character string with the name of the variable unique for each tree.}

\item{dbh}{a character string with the name of the variable with the diameter of the trees}

\item{mscale}{a character indicating the mesured scale (mm or cm).}
}
\value{
a data frame with the same columns as plotdata plus a 'dbh.cm' column.
}
\description{
Functions to estimate a single dbh from a multiple stem tree or shrub.
}
\section{Details}{


Calculate the sum of basal areas for multiple stem tree and them return the dbh associated with the sum of basala area.
}

\section{References}{

}

\examples{

\dontrun{
data(bci7)
plotData <- singleStem(bci7, mscale= "mm")
}
}
